More granular API functions
This is as requested in #575

The idea here is to less "black box" and also easy to use.
There are separate Deployer and Updater interfaces with functions
that you call in sequence. To see how MSFT would use it look at the new
implementation of CreateOrUpdate. They would just use their own implementation
 instead of DefaultDeployment.